:recycle: refactor(`<ConsultationsPage>`): Extract `makeConsultationsPage`.
