+++ imports/ui/consultations/ConsultationsPage.ts
@@ -1,0 +1,9 @@
+import makeConsultationsPage from './makeConsultationsPage';
+
+import useConsultationsAndAppointments from './useConsultationsAndAppointments';
+
+const ConsultationsPage = makeConsultationsPage(
+	useConsultationsAndAppointments,
+);
+
+export default ConsultationsPage;
